+++ src/main/java/racingcar/controller/CarRaceGame.java
@@ -5,14 +5,15 @@ import java.util.Arrays;
 import java.util.List;
 import racingcar.domain.Car;
 import racingcar.domain.GameManager;
-import racingcar.utility.CarRaceGameUtility;
+import racingcar.utility.GameExecutionUtility;
+import racingcar.utility.GameSettingUtility;
 import racingcar.validator.Validator;
 import racingcar.view.CarRaceGameView;
 import static camp.nextstep.edu.missionutils.Console.*;
 
 public class CarRaceGame {
     private GameManager gameManager;
-    private CarRaceGameUtility carRaceGameUtility;
+    private GameExecutionUtility carRaceGameUtility;
 
     private CarRaceGame() {
     }
@@ -22,52 +23,18 @@ public class CarRaceGame {
     }
 
     public static void startGame() {
-        List<String> carNameList = getCarNameListAndValidate();
-        setCarImpls(carNameList);
+        List<String> carNameList = GameSettingUtility.getCarNameListAndValidate();
+        GameSettingUtility.setCarImpls(carNameList);
 
-        CarRaceGameUtility.executeCarRaceGameForNRound(getGameRoundAndValidate());
+        GameExecutionUtility.executeCarRaceGameForNRound(GameSettingUtility.getGameRoundAndValidate());
 
-        List<String> winnerList = CarRaceGameUtility.findWinner(GameManager.getCarImplList());
+        List<String> winnerList = GameExecutionUtility.findWinner(GameManager.getCarImplList());
         CarRaceGameView.printGameResult(winnerList);
 
         endGame();
     }
 
-    public static List<String> getCarNameListAndValidate() {
-        CarRaceGameView.printStartGame();
-        String carNames = CarRaceGameView.getInput();
-        Validator.checkCarNamesForm(carNames);
 
-        List<String> carNameList = Arrays.stream(carNames.split(",")).toList();
-        validateCarName(carNameList);
-        return carNameList;
-    }
-
-    public static int getGameRoundAndValidate() {
-        CarRaceGameView.printGameRoundNumberToRun();
-        String gameRoundNumberString = CarRaceGameView.getInput();
-        validateGameRoundNumber(gameRoundNumberString);
-        return Integer.parseInt(gameRoundNumberString);
-    }
-
-    private static void setCarImpls(List<String> carNameList) {
-        carNameList.forEach(carName -> GameManager.addCarImpl(new Car(carName)));
-    }
-
-
-    public static void validateCarName(List<String> carNameList) {
-        Validator.checkCarNameDuplication(carNameList);
-        for (String carName : carNameList) {
-            Validator.isLessThanFiveLetter(carName);
-            Validator.isEmptyString(carName);
-        }
-    }
-
-    public static void validateGameRoundNumber(String gameRound) {
-        Validator.isEmptyString(gameRound);
-        Validator.isNumber(gameRound);
-        Validator.isMoreThanOne(gameRound);
-    }
 
 
     public static void endGame() {
